# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "SwaggerHUB API Mocking"
        }
    ],
    "paths": {
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "All Categories",
                "description": "Multiple status values can be provided with comma separated string",
                "operationId": "indexCategories",
                "parameters": [
                    {
                        "name": "products",
                        "in": "query",
                        "description": "All categories",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "available",
                            "enum": [
                                "available",
                                "pending",
                                "sold"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/categories/1": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get category by id",
                "description": "Multiple status values can be provided with comma separated string",
                "operationId": "showCategories",
                "parameters": [
                    {
                        "name": "categoris",
                        "in": "query",
                        "description": "Get category by id",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "available",
                            "enum": [
                                "available",
                                "pending",
                                "sold"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/categories/store": {
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Create categories",
                "operationId": "storeCategories",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Enter name of the category",
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "Enter category image",
                                        "type": "file"
                                    },
                                    "title": {
                                        "description": "Enter category title",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Enter category description",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/categories/update/1": {
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Update category",
                "operationId": "updateCategories",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Enter name of the category",
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "Enter category image",
                                        "type": "file"
                                    },
                                    "title": {
                                        "description": "Enter category title",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Enter category description",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/categories/destroy/1": {
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "delete category",
                "operationId": "destroyCategories",
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Sanctum security laravel",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}